requirements.txt
bcrypt==3.2.0
cachelib==0.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Paranoid==0.2.0
Flask-Session==0.4.0
flask-sock==0.5.1
Flask-SQLAlchemy==2.5.1
h11==0.13.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
msal==1.17.0
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
simple-websocket==0.5.1
six==1.16.0
SQLAlchemy==1.4.32
Turbo-Flask==0.8.0
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.1.0
gunicorn==20.1.0
python-dateutil==2.8.2